backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
fastapi==0.115.6
filelock==3.20.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
packaging==25.0
pillow==10.4.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.12
PyYAML==6.0.3
safetensors==0.7.0
shellingham==1.5.4
starlette==0.41.3
sympy==1.14.0
timm==1.0.9
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.32.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
